LPS-102897 Replace files, removing the extra one
